Working on voxel addition/deletion

+++ Lynx/src/Lynx/Render/Buffer.cpp
@@ -80,7 +80,7 @@ namespace Lynx {
 		glBindBuffer(GL_ARRAY_BUFFER, 0);
 	}
 
-	void VertexBuffer::SetData(const void* data, uint32_t size)
+	void VertexBuffer::SetData(const void* data, size_t size)
 	{
 		glBindBuffer(GL_ARRAY_BUFFER, m_VertexBufferID);
 		glBufferSubData(GL_ARRAY_BUFFER, 0, size, data);
