+++ Lynx/src/Lynx/Render/Buffer.h
@@ -71,7 +71,7 @@ namespace Lynx {
 		~VertexBuffer();
 		void Bind() const;
 		void Unbind() const;
-		void SetData(const void* data, uint32_t size);
+		void SetData(const void* data, size_t size);
 
 		const BufferLayout& GetLayout() const { return m_Layout; }
 		void SetLayout(const BufferLayout& layout) { m_Layout = layout; }
